+++ src/org/ods/component/RolloutOpenShiftDeploymentStage.groovy
@@ -5,6 +5,7 @@ import groovy.transform.TypeCheckingMode
 import org.ods.services.OpenShiftService
 import org.ods.services.JenkinsService
 import org.ods.util.ILogger
+import org.ods.openshift.OpenShiftResourceMetadata
 import org.ods.util.PodData
 
 @SuppressWarnings('ParameterCount')
@@ -129,6 +130,8 @@ class RolloutOpenShiftDeploymentStage extends Stage {
             tailorApply(context.targetProject)
             refreshResources = true
         }
+        def metadata = new OpenShiftResourceMetadata(steps, context.properties, options.properties, logger, openShift)
+        metadata.updateMetadata()
 
         if (refreshResources) {
             deploymentResources = openShift.getResourcesForComponent(